phase1/requirements.txt
appnope==0.1.3
asttokens==2.3.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.3
fonttools==4.42.1
idna==3.4
imageio==2.31.2
ipympl==0.9.3
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
widgetsnbextension==4.0.8
